+++ NEWS.md
@@ -1,6 +1,15 @@
-# visibly 0.1.0
+## visibly 0.2.0
 
-* Added a `NEWS.md` file to track changes to the package.
+* Package website.
+
+## visibly 0.1.5
+
+* Add coefficient plots.
+
+## visibly 0.1.0
+
+* Initial start of package.
+